Updating sh_tr.lua and fixing some translate error
I noticed that the previous translation was done using a machine translation tool like Google Translate, and I wanted to provide a professional translation for your add-on. If there are any errors, please feel free to send me a message on Gmodstore or submit a pull request for a revision.

I have included my Gmodstore link in the last line. Best regards.